add use_mdfile_as_mainpage option
This makes it possible/easier to use a markdown file as mainpage, e.g. the README.md

To use it, add a rosdoc.yaml with
```
- builder: doxygen
  file_patterns: '*.c *.cpp *.h *.cc *.hh *.dox *.md'
  use_mdfile_as_mainpage: README.md
```

Description of this option from Doxygen:

If the USE_MDFILE_AS_MAINPAGE tag refers to the name of a markdown file that
is part of the input, its contents will be placed on the main page
(index.html). This can be useful if you have a project on for instance GitHub
and want to reuse the introduction page also for the doxygen output.